- added the changes according to the comments - added package.json
