+++ package.json
@@ -15,6 +15,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.21.1",
+    "knex": "^3.1.0",
     "pg": "^8.13.1"
   },
   "devDependencies": {
