+++ repositories/roles_repository.js
@@ -1,9 +1,11 @@
 const { getKnex } = require("../db");
- 
-
 async function getAllRoles() {
     const knex = getKnex();
     return await knex('roles').select('*');
-
 }
-module.exports = { getAllRoles };
+async function assignRoleToUser(user_id, role_id) {
+    const knex = getKnex();
+    await knex("user_roles").insert({ role_id, user_id });
+}
+
+module.exports = { getAllRoles, assignRoleToUser };
