+++ server.js
@@ -2,18 +2,16 @@
 require("dotenv").config();
 const express = require("express");
 const cors = require("cors");
-
 const app = express();
 const { initDb } = require('./db');
-const rolesService = require('./services/roles_service');
+const utils = require('./utils');
 // Middleware
 app.use(express.json()); // read documentation on what this does
 app.use(cors());
 
 (async () => { 
-    await initDb();
-    console.log("1");
-  await rolesService.loadRolesIntoMemory();
+  await initDb();
+  await utils.loadRolesIntoMemory();
 
   // Routes
   const routes = require("./routes/routes");
